ProActive

`ProActive` automatically detects regions of gapped and elevated read coverage using a 2D pattern-matching algorithm. `ProActive` detects, characterizes and visualizes read coverage patterns in both genomes and metagenomes. Optionally, users may provide gene annotations associated with their genome or metagenome in the form of a .gff file. In this case, `ProActive` will generate an additional output table containing the gene annotations found within the detected regions of gapped and elevated read coverage. Additionally, users can search for gene annotations of interest in the output read coverage plots.

Details

The three main functions in `ProActive` are:

ProActiveDetect performs the pattern-matching and characterization of read coverage patterns.

plotProActiveResults plots the results from ProActiveDetect()

geneAnnotationSearch searches classified contigs/chunks for gene annotations that match user-provided keywords.

See also

Author

Jessie Maier jlmaier@ncsu.edu