Change the pileup window size

Re-averages windows of pileup files with 100bp windows to reduce pileup size.

Usage

changewindowSize(pileupSubset, windowSize, mode)

Arguments

pileupSubset

A subset of the pileup that pertains only to the contig/chunk currently being assessed.

windowSize

The number of basepairs to average read coverage values over. Options are 100, 200, 500, 1000 ONLY. Default is 1000.

mode

Either "genome" or "metagenome"