Shrink the width of full elevation and gap patterns

Remove values from gapped/elevated region in the pattern-match vector until it reaches the `minSize`

Usage

fullElevGapShrink(
  minCov,
  windowSize,
  maxCov,
  elevLength,
  nonElev,
  bestMatchInfo,
  pileupSubset,
  minSize,
  elevOrGap
)

Arguments

minCov

The minimum value of the pattern-match vector.

windowSize

The number of basepairs to average read coverage values over. Options are 100, 200, 500, 1000 ONLY. Default is 1000.

maxCov

The maximum value of the pattern-match vector.

elevLength

Length of the elevated/gapped pattern-match region.

nonElev

Length of the non-elevated/gapped pattern-match region.

bestMatchInfo

The information associated with the current best pattern-match for the contig/chunk being assessed.

pileupSubset

A subset of the pileup that pertains only to the contig/chunk currently being assessed.

minSize

The minimum size (in bp) of elevation or gap patterns. Default is 10000.

elevOrGap

Pattern-matching on 'elevation' or 'gap' pattern.