sampleMetagenomePileup

A subset of contigs from the raw whole-community fraction read coverage pileup file generated during read mapping. Report...

Usage

sampleMetagenomePileup

Format

## `sampleMetagenomePileup` A data frame with 4,604 rows and 4 columns:

V1

Contig accession

V2

Mapped read coverage averaged over a 100 bp window size

V3

Starting position (bp) of each 100 bp window. Restarts from 0 at the start of each new contig.

V4

Starting position (bp) of each 100 bp window. Does NOT restart at the start of each new contig.

Source

<https://microbiomejournal.biomedcentral.com/articles/10.1186/s40168-020-00935-5>

Details

This dataset was generated from a conventional mouse fecal homogenate. The whole-community extracted DNA was sequenced with Illumina (paired-end mode, 150 bp reads) after which the metagenome was assembled. The sequencing reads were mapped to the assembled contigs using BBMap. The bbmap.sh bincov parameter with covbinsize=100 was used to create a pileup file with 100 bp windows. A subset of 10 contigs from the pileup file were selected for this sample dataset. The contigs were chosen because their associated read coverage patterns exemplify ProActive's pattern-matching and characterization functionality across classifications: NODE_1911: elevation off left NODE_1583: elevation off right NODE_1884: gap off right NODE_1255: gap off left NODE_368: full gap NODE_617: elevation full NODE_1625: no pattern